[FIX] performance issue with module loading in new API
Old API made a difference between falsy values at the storage level ('' would
be stored as '' and False as False), but client itself would not, and new API
changes behavior to take after client and unify handling. This means a write
of `''` in the old API would be read as `''`, but in the one it'll be read as
``False`` breaking code relying on such a behavior for eq. Such as the
update_list code checking if a module's manifest has been updated since the
last load-to-db.

Fix by special-casing situations where both old and new values are falsy.