[IMP] comment out IdType assertions in BaseModel.browse
These turn out to have significant overhead, especially on e.g. new API-based
search-heavy code, the gencomp is a multiplier and may lead to million+ calls
to isinstance with a class tuple, contributing significantly to the runtime of
the system.

Commenting that lowers the performance worsening of the new API when
installing account to ~5%.

If we want to re-enable it for security purposes, "internal browsing"
(e.g. from search) should use the assertionless _browse instead.