[FIX] fields.Binary read: restore compatibility with old bin_size API
We used to return the "bin_size" version of a binary field
using a human-readable format with both a value and a unit
of measure. Some heuristics in the web client depend on it
and we can easily mimic that behavior using the new
implementation.
In addition we can continue to support the more specific
`bin_size_<field>` context key to selectively turn
bin_size on per-field.

Finally, the new implementation depends on the
pg_size_pretty() PostgresSQL function that changed in
PG 9.2 in a somewhat incompatible manner. So we force
the ::bigint cast to avoid errors in 9.2+.
See also:
http://postgresql.1045698.n5.nabble.com/No-pg-size-pretty-numeric-was-not-such-a-hot-idea-td5710106.html