[IMP] micro-optimization: dict.pop -> key in dict + del dict[key]
Note: all measurements are from my machine, but the relative gain should be
valid for any CPython (hopefully)

* if the key is in the dict, ``dict.pop(key, default)`` and
  ``if key in dict: del dict[key]`` have about the same runtime.
* otoh if the key is *not* in the dict, the latter is ~3 times faster (0.06
  usec to 0.183 usec on my machine).

Because invalidate is a tight loop where 99% of the iterations of the fields
are *not* in the cache during module installation (for account, I get 19
million checks and 126000 removals), moving from pop to if+del yields a 10%
runtime gain, from 75s to 67s.

An even better optimization would be some sort of algorithmic method to avoid
the 20m iteration/test/test: although they only take 1 usec each, they add up
to a significant runtime.