[FIX] improve average runtime of BaseModel.modified
During loading of XML data files during the installation of ``account``, it
turns out the *vast* majority of non-stored function fields are not cached,
and asking for their invalidation leads to a huge runtime hit for "nothing"
(spinning wheels in ``Environment.invalidate`` but ultimately not doing
anything since the fields aren't cached).

Checking whether the fields should be invalidated before adding them to the
specs list (by collecting all currently cached fields in a set and only
selecting fields for invalidation if they are in the set) leads to ~6x less
hits within invalidate, and lowers the average runtime of BaseModel.modified
from 49s to 18s on my machine.