[REV] a Bundle in the Head
Not a Rage Against the Machine song but a revert in order to place
the javascript bundles back in the <head>.

In v8.0 the javascript bundles were moved in the bottom of the page for
SEO reasons. This was done in commit 5c87e72.
This change pointed out that there were still some leftover in addons
that should be bundlified. Those were fixed in 8d08e0.

For v9.0, the javascript bundles will be brought back to the <head> in
order to allow users to monkey patch libs or customize js stuff.