[FIX] web_tip: gc tip pipeline on view manager switch mode
When more than one tip are in the tip pipeline for a view, navigating
away by consuming a tip (e.g. by clicking on the highlighted breadcrumb)
can cause the following tip to be displayed anyway, breaking the ui.

This commit adds a gc method that cleans the remaining tips. It's called
in the overriden ViewManager's switch_mode method.