[FIX] set correct id on default searchviews (web client)
The problem is that when a default searchview is requested, it is
initialized without its view_id.  But when the fields_view_get is
received, it knows its view_id.  This is not a really big problem,
but it prevents debug mode from editing the view.  This patch solves
the issue.