[FIX] product.uom: remove unnecessary rounding step, add missing test
Remove the intermediate rounding inside _compute_qty(), as it
is not necessary after rev.fa2f7b86 and has undesired sife-effects.

An extra float_round() operation inside _compute_qty
had been added at rev. 311c77bb to avoid a float representation
error in UoM factors to bias the ceiling() operation
done as the last conversion step.

Example 1:
Dozen has a factor of 1/12, which was previously stored in the
database with a decimal accuracy of 12 significant decimal digits.
This meant the factor was exactly stored as 0.08333333333333.
When reading this back into a Python float, the precision was not
sufficient, and the UoM conversion of 1 Dozen to Units gave a
result of 12.00000000000047961...
After the final ceiling() operation to Unit's rounding, the
converted value ended up as 13.

This is the problem that was solved using an extra rounding.

However at revision fa2f7b86 the decimal precision used to store
UoM factors was increased to preserve all significant digits.
This added the extra precision necessary to read the Dozen factor
back into an accurate float value of 1/12, and the conversion of
1 Dozen now gives 12.0 Units, even without the intermediate
rounding operation.

On the other hand that extra rounding operation has undesired
side-effects, as it requires a fixed precision derived from
the rounding precision of UoMs, while there is no given
precision that would work in all cases for this intermediate
value. It would always be possible to find a valid combination
of UoM roundings that breaks it.

Example 2:
Let Grams have a rounding precision set to 1 because no smaller
quantities are allowed, and Kilograms a rounding of 0.001 to allow
representing 1 Gram. (gram factor = 1000 by default)
If we try to convert 1234 Grams into Kilograms, the extra rounding
introduced in 311c77bb will cause a rounding of 1234.0/1000.0 at
the precision of Grams (1), which gives 1.0 as a result.
The net result of this conversion gives 1234.0 Gram = 1.0 Kilogram,
while the correct result of 1.234 Kilogram is perfectly compatible
with the settings.

Similar errors could be triggered with various rounding settings, as
long as the intermediate rounding uses a fixed precision.

Two extra tests have been added to cover Example 1 and Example 2.