[FIX] product_extended: incorrect inheritance
This view was intended to extend the tree view of mrp.bom, not
mrp.bom.line. Nonetheless, the view mrp.mrp_bom_tree_view which is extended here is a
view for a mrp.bom.line, and there is no tree view for a mrp.bom.

So this extension has no meaning anymore and seems to have worked only
by accident...