[FIX] models, fields: reorganize the fields setup to not miss dependencies
The fields setup was sometimes missing a trigger on a dependency.  Suppose a
field depends on a one2many; you should add a trigger on its inverse field
(many2one), but the latter is unknown if the one2many has not been set up yet!
As a consequence, modifying the many2one was not triggering the recomputation
of the field.

To avoid this bug, the setup has been split in three phases:
(1) determine all inherited and custom fields on models;
(2) setup fields, except for recomputation triggers;
(3) add recomputation triggers and complete the setup of the model.

Making these three phases explicit brings good invariants:
- when setting up a field, you know that models already know all their field;
- when adding recomputation triggers, you know that fields have been set up.